Remove page view tracking.
When page view tracking was added, a unique cookie was associated with
each session, and tied to a user during authentication. For unexplored
reasons, the cookie value was sometimes corrupted causing an error when
used during sign in. Since users don't care about being tracked, and
nobody looks at the analytics anyway, I'm choosing to remove this to fix
the sign-in bug, instead of spending the time to fix the page tracking.

The proper fix would be to implement UUID validation of tracking cookies
*before* they are sent to the server, and have the server ignore any
invalid tracking cookies.