Replace report form on comment page with one-click report link on comments.
After I spent time merging and finishing the work Dallas started, I
noticed that the spec calls for a modal popup not a separate page. The
navigation links on the comment page were also removed, and the comment
page has been turned into the a petition page. With that considered, I
decided a better UX was to just have a one-click "report" link with no
explanation.

If it's important to include an explanation, we need to either revisit
the spec or implement it according to spec.