Fix potential call to undefined function.
I inlined the function that dedupes fetched comments because the other
functions on the class refered to it using "this" but when those
functions are called from other components it would be missing.

This is another reason I want to finish my refactor to use single state functional architecture (raj library) rather
than the react-ish class hyperloop architecture with a bunch of magic and component state.